Disabled index pattern checks
