+++ public/react-services/pattern-handler.js
@@ -28,42 +28,6 @@ export class PatternHandler {
       const selectedPattern = AppState.getCurrentPattern();
       if (selectedPattern && selectedPattern !== pattern) defaultPatterns.push(selectedPattern);
       let patternList = await SavedObject.getListOfWazuhValidIndexPatterns(defaultPatterns, origin);
-
-      if (origin === HEALTH_CHECK) {
-        const indexPatternFound = patternList.find((indexPattern) => indexPattern.title === pattern);
-
-        if (!indexPatternFound && pattern) {
-          // if no valid index patterns are found we try to create the wazuh-alerts-*
-          try {
-            getToasts().add({
-              color: 'warning',
-              title:
-                `No ${pattern} index pattern was found, proceeding to create it.`,
-              toastLifeTimeMs: 5000
-            });
-
-            if (await SavedObject.getExistingIndexPattern(pattern)) {
-              await SavedObject.refreshIndexPattern(pattern);
-              getToasts().addSuccess(`${pattern} index pattern updated successfully`);
-            } else {
-              await SavedObject.createWazuhIndexPattern(pattern);
-              getToasts().addSuccess(`${pattern} index pattern created successfully`);
-            }
-
-            patternList = await SavedObject.getListOfWazuhValidIndexPatterns(defaultPatterns, origin);
-            !AppState.getCurrentPattern() && AppState.setCurrentPattern(pattern);
-          } catch (err) {
-            getToasts().addDanger({
-              title: 'Error creating the index pattern.',
-              text: err.message || err,
-              toastLifeTimeMs: 3000
-            });
-            AppState.removeCurrentPattern();
-
-            return;
-          }
-        }
-      }
       if (AppState.getCurrentPattern() && patternList.length) {
         let filtered = patternList.filter(
           item => item.id === AppState.getCurrentPattern()
